split logic out, so one can run it with Github Actions (which doesn't accept CLI arguments) or as a CLI program (which accepts arguments) with node dist/cli.js <filename>
